requirements.txt
asgiref==3.8.1
attrs==24.3.0
Brotli==1.1.0
cffi==1.17.1
cssselect2==0.7.0
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
fonttools==4.55.3
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pillow==11.1.0
psycopg==3.2.4
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
uritemplate==4.1.1
weasyprint==63.1
webencodings==0.5.1
zopfli==0.2.3.post1
